- refactor: 세션 유저 조회 로직 SessionUserProvider로 통일 - body: - SessionUserProvider 도입 및 OrderService/CartService/OrderQueryService 적용 - userId Long 정규화 및 401 예외 흐름 일원화

+++ src/main/java/com/jkomerce/store/controller/BoardController.java
@@ -34,7 +34,7 @@ public class BoardController {
     }
 
     @PutMapping("/{id}")
-    public int modifyBoard(@PathVariable int id, @RequestBody BoardDTO boardDTO) {
+    public int modifyBoard(@PathVariable Long id, @RequestBody BoardDTO boardDTO) {
 
         boardDTO.setId(id);
         return boardService.updateBoard(boardDTO);
